# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
...
  domain.CreateProductRequest:
    properties:
      berat:
        minimum: 0
        type: integer
      deskripsi:
        type: string
      harga_konsumen:
        minimum: 0
        type: number
      harga_reseller:
        minimum: 0
        type: number
      id_category:
        type: integer
      nama_produk:
        maxLength: 255
        minLength: 2
        type: string
      stok:
        minimum: 0
        type: integer
    required:
    - harga_konsumen
    - harga_reseller
    - id_category
    - nama_produk
    type: object
...
  domain.PhotoProduk:
    properties:
      created_at:
        type: string
      id:
        type: integer
      id_produk:
        type: integer
      is_primary:
        type: boolean
      position:
        type: integer
      product:
        allOf:
        - $ref: '#/definitions/domain.Product'
        description: Relations
      updated_at:
        type: string
      url:
        type: string
    type: object
  domain.Product:
    properties:
      berat:
        type: integer
      category:
        $ref: '#/definitions/domain.Category'
      created_at:
        type: string
      deskripsi:
        type: string
      harga_konsumen:
        minimum: 0
        type: number
      harga_reseller:
        minimum: 0
        type: number
      id:
        type: integer
      id_category:
        type: integer
      id_toko:
        type: integer
      nama_produk:
        maxLength: 255
        minLength: 2
        type: string
      photos:
        items:
          $ref: '#/definitions/domain.PhotoProduk'
        type: array
      slug:
        type: string
      sold_count:
        type: integer
      status:
        type: string
      stok:
        type: integer
      toko:
        allOf:
        - $ref: '#/definitions/domain.Store'
        description: Relations
      updated_at:
        type: string
    required:
    - harga_konsumen
    - harga_reseller
    - nama_produk
    type: object
...
  domain.Store:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      photo_url:
        type: string
      rating:
        type: number
      status:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/domain.User'
        description: Relations
      user_id:
        type: integer
    required:
    - name
    type: object
...
  domain.UpdateProductRequest:
    properties:
      berat:
        minimum: 0
        type: integer
      deskripsi:
        type: string
      harga_konsumen:
        minimum: 0
        type: number
      harga_reseller:
        minimum: 0
        type: number
      id_category:
        type: integer
      nama_produk:
        maxLength: 255
        minLength: 2
        type: string
      stok:
        minimum: 0
        type: integer
    required:
    - harga_konsumen
    - harga_reseller
    - id_category
    - nama_produk
    type: object
...
  domain.User:
    properties:
      about:
        type: string
      city_id:
        type: integer
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      gender:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      job:
        type: string
      last_login_at:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      photo_url:
        type: string
      province_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    - phone
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: E-commerce API built with Go Fiber and Clean Architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Commerce API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Get all products with pagination and filtering (public endpoint)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by product name
        in: query
        name: search
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: string
      - default: newest
        description: 'Sort by: newest, oldest, price_asc, price_desc'
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Product'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product for the authenticated user's store
      parameters:
      - description: Product creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Product'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing product (owner only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - not product owner
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get a single product by its ID (public endpoint)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Product'
              type: object
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product (owner only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Product'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - not product owner
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
  /products/{id}/photos:
    post:
      consumes:
      - multipart/form-data
      description: Upload a photo for a product (owner only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Photo file (JPG, JPEG, PNG, max 5MB)
        in: formData
        name: photo
        required: true
        type: file
      - default: false
        description: Set as primary photo
        in: formData
        name: is_primary
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Photo uploaded successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.PhotoProduk'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - not product owner
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Upload product photo
      tags:
      - Products
  /products/{id}/photos/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete a photo from a product (owner only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Photo deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - not product owner
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete product photo
      tags:
      - Products
  /products/{id}/photos/{photoId}/primary:
    put:
      consumes:
      - application/json
      description: Set a photo as the primary photo for a product (owner only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Primary photo set successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - not product owner
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Set primary photo
      tags:
      - Products
  /products/my:
    get:
      consumes:
      - application/json
      description: Get all products owned by the authenticated user with pagination
        and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by product name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Product'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current user's products
      tags:
      - Products
  /products/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get a single product by its slug (public endpoint)
      parameters:
      - description: Product slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Product'
              type: object
        "400":
          description: Invalid slug
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get product by slug
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"